extract health handler into function
