+++ src/cmd/delegation_backend/main_bpu.go
@@ -3,7 +3,6 @@ package main
 import (
 	. "block_producers_uptime/delegation_backend"
 	"context"
-	"encoding/json"
 	"net/http"
 	"time"
 
@@ -14,11 +13,6 @@ import (
 	"google.golang.org/api/option"
 	sheets "google.golang.org/api/sheets/v4"
 )
-
-// HealthStatus represents the JSON response structure for the /health endpoint
-type HealthStatus struct {
-	Status string `json:"status"`
-}
 
 func main() {
 	// Setup logging
@@ -126,15 +120,9 @@ func main() {
 	http.Handle("/v1/submit", app.NewSubmitH())
 
 	// Health check endpoint
-	http.HandleFunc("/health", func(rw http.ResponseWriter, r *http.Request) {
-		if app.IsReady {
-			rw.WriteHeader(http.StatusOK)
-			json.NewEncoder(rw).Encode(HealthStatus{Status: "ok"})
-		} else {
-			rw.WriteHeader(http.StatusServiceUnavailable)
-			json.NewEncoder(rw).Encode(HealthStatus{Status: "unavailable"})
-		}
-	})
+	http.HandleFunc("/health", HealthHandler(func() bool {
+		return app.IsReady
+	}))
 
 	// Sheets service and whitelist loop
 	app.WhitelistDisabled = appCfg.DelegationWhitelistDisabled
